added missing properties for type
