+++ components/availability/src/Availability.svelte
@@ -19,6 +19,7 @@
   import {
     getEventDispatcher,
     buildInternalProps,
+    arrayContainsArray,
   } from "@commons/methods/component";
   import type { TimeInterval } from "d3-time";
   import { timeWeek, timeDay, timeHour, timeMinute } from "d3-time";
@@ -61,6 +62,7 @@
   import { createConsecutiveQueryKey } from "./method/consecutive";
   import type { EventDefinition } from "@commonstypes/ScheduleEditor";
   import type { ConsecutiveEvent } from "@commonstypes/Scheduler";
+  import { generateDaySlots, overlap } from "./method/availabilityUtils";
 
   //#region props
   export let id: string = "";
@@ -368,230 +370,6 @@
 
   $: endDay = dayRange[dayRange.length - 1];
 
-  const allRequiredParticipantsIncluded = (calendars: string[]) => {
-    return requiredParticipants.every((participant) =>
-      calendars.includes(participant),
-    );
-  };
-
-  // map over the ticks() of the time scale between your start day and end day
-  // populate them with as many slots as your start_hour, end_hour, and slot_size dictate
-  $: generateDaySlots = function (
-    timestamp: Date,
-    startHour: number,
-    endHour: number,
-  ) {
-    const dayStart = timeHour(
-      new Date(new Date(timestamp).setHours(startHour)),
-    );
-    const dayEnd = timeHour(new Date(new Date(timestamp).setHours(endHour)));
-    return scaleTime()
-      .domain([dayStart, dayEnd])
-      .ticks(timeMinute.every(_this.slot_size) as TimeInterval)
-      .slice(0, -1) // dont show the 25th hour
-      .map((time: Date) => {
-        const endTime = timeMinute.offset(time, _this.slot_size);
-        const freeCalendars: string[] = [];
-        let availability = AvailabilityStatus.FREE; // default
-        if (allCalendars.length) {
-          for (const calendar of allCalendars) {
-            // Adjust calendar.timeslots for buffers
-            const timeslots =
-              calendar.availability === AvailabilityStatus.BUSY
-                ? calendar.timeslots.map((slot: TimeSlot) => ({
-                    start_time: timeMinute.offset(
-                      slot.start_time,
-                      -_this.event_buffer,
-                    ),
-                    end_time: timeMinute.offset(
-                      slot.end_time,
-                      _this.event_buffer,
-                    ),
-                    available_calendars: slot.available_calendars,
-                  }))
-                : calendar.timeslots.map((slot: TimeSlot) => ({
-                    // Don't apply start-buffer to the first timeslot, nor end-buffer to the last timeslot.
-                    // Works across multiple days; you won't get a random buffer at 11:50 / 00:10
-                    start_time: timeMinute.offset(
-                      slot.start_time,
-                      slot === calendar.timeslots[0] ? 0 : _this.event_buffer,
-                    ),
-                    end_time: timeMinute.offset(
-                      slot.end_time,
-                      slot === calendar.timeslots[calendar.timeslots.length - 1]
-                        ? 0
-                        : -_this.event_buffer,
-                    ),
-                    available_calendars: slot.available_calendars,
-                  }));
-
-            const slot: TimeSlot = {
-              start_time: time,
-              end_time: endTime,
-              available_calendars: [],
-            };
-
-            let concurrentSlotEventsForUser = 0;
-            if (calendar.availability === AvailabilityStatus.BUSY) {
-              // For Busy calendars, a timeslot is considered available if its calendar has no overlapping events
-              concurrentSlotEventsForUser = overlap(timeslots, slot);
-            } else if (calendar.availability === AvailabilityStatus.FREE) {
-              // For Free calendars, a timeslot is considered available if its calendar has a time that fully envelops it.
-              concurrentSlotEventsForUser = timeslots.some(
-                (blob: TimeSlot) =>
-                  slot.start_time >= blob.start_time &&
-                  slot.end_time <= blob.end_time,
-              )
-                ? 1
-                : 0;
-              // slot availability is when a given timeslot has all of its minutes represented in calendar.free timeslots
-            }
-            if (calendar.availability === AvailabilityStatus.BUSY) {
-              if (
-                _this.capacity &&
-                _this.capacity >= 1 &&
-                concurrentSlotEventsForUser < _this.capacity
-              ) {
-                freeCalendars.push(calendar?.account?.emailAddress || "");
-              } else if (!concurrentSlotEventsForUser) {
-                freeCalendars.push(calendar?.account?.emailAddress || "");
-              }
-            } else if (
-              calendar.availability === AvailabilityStatus.FREE ||
-              !calendar.availability
-            ) {
-              // if a calendar is passed in without availability, assume the timeslots are available.
-              if (concurrentSlotEventsForUser) {
-                freeCalendars.push(calendar?.account?.emailAddress || "");
-              }
-            }
-          }
-
-          if (freeCalendars.length) {
-            if (freeCalendars.length === allCalendars.length) {
-              availability = AvailabilityStatus.FREE;
-            } else {
-              availability = AvailabilityStatus.PARTIAL;
-            }
-          } else {
-            availability = AvailabilityStatus.BUSY;
-          }
-        }
-
-        // If availability is partial (not 100% of calendars), and that ratio is less than partial_bookable_ratio,
-        // mark the slot as busy
-        if (
-          availability === AvailabilityStatus.PARTIAL &&
-          freeCalendars.length <
-            allCalendars.length * _this.partial_bookable_ratio
-        ) {
-          availability = AvailabilityStatus.BUSY;
-        }
-
-        // Allows users to book over busy slots if partial_bookable_ratio is 0
-        if (
-          availability === AvailabilityStatus.BUSY &&
-          _this.partial_bookable_ratio === 0
-        ) {
-          availability = AvailabilityStatus.PARTIAL;
-        }
-
-        // If availability is partial, but a required participant is unavailble, the slot becomes Busy
-        if (
-          availability === AvailabilityStatus.PARTIAL &&
-          _this.required_participants.length
-        ) {
-          if (!allRequiredParticipantsIncluded(freeCalendars)) {
-            availability = AvailabilityStatus.BUSY;
-          }
-        }
-
-        // If mandate_top_of_hour, change any status to "busy" if it's not at :00
-        if (_this.mandate_top_of_hour && time.getMinutes() !== 0) {
-          availability = AvailabilityStatus.BUSY;
-          freeCalendars.length = 0;
-        }
-
-        // if the "open_hours" property has rules, adhere to them above any other event-based free/busy statuses
-        // (Mark the slot busy if it falls outside the open_hours)
-        if (_this.open_hours.length) {
-          if (availability !== AvailabilityStatus.BUSY) {
-            let dayRelevantRules = [];
-            dayRelevantRules = _this.open_hours.filter(
-              (rule) =>
-                !rule.hasOwnProperty("startWeekday") ||
-                rule.startWeekday === time.getDay() ||
-                rule.endWeekday === time.getDay(),
-            );
-            let slotExistsInOpenHours = dayRelevantRules.some((rule, iter) => {
-              let ruleStartAppliedDate = rule.hasOwnProperty("startWeekday")
-                ? timeDay.offset(
-                    time,
-                    (rule.startWeekday as number) - time.getDay(),
-                  )
-                : new Date(time);
-              ruleStartAppliedDate.setHours(rule.startHour);
-              ruleStartAppliedDate.setMinutes(rule.startMinute);
-
-              let ruleEndAppliedDate = rule.hasOwnProperty("startWeekday")
-                ? timeDay.offset(
-                    time,
-                    (rule.endWeekday as number) - time.getDay(),
-                  )
-                : new Date(time);
-              ruleEndAppliedDate.setHours(rule.endHour);
-              ruleEndAppliedDate.setMinutes(rule.endMinute);
-
-              return (
-                time >= ruleStartAppliedDate && endTime <= ruleEndAppliedDate
-              );
-            });
-            if (!slotExistsInOpenHours) {
-              availability = AvailabilityStatus.CLOSED;
-              freeCalendars.length = 0;
-            }
-          }
-        }
-
-        // Handle the Consecutive Events model, where a slot if busy if it falls within a consecutive timeslot.
-        // None of the other above rules should apply, except (maybe!) Buffer and Open Hours.
-        if (_this.events?.length > 1 && consecutiveOptions.length) {
-          const existsWithinConsecutiveBlock = consecutiveOptions.some(
-            (event) => {
-              return (
-                time >= event[0].start_time &&
-                endTime <= event[event.length - 1].end_time
-              );
-            },
-          );
-          if (existsWithinConsecutiveBlock) {
-            availability = AvailabilityStatus.FREE;
-            freeCalendars.length = consecutiveParticipants.length;
-          } else {
-            availability = AvailabilityStatus.BUSY;
-            freeCalendars.length = 0;
-          }
-        }
-
-        const dateOffset =
-          (new Date(time).getTime() - new Date().getTime()) /
-          (1000 * 60 * 60 * 24);
-
-        return {
-          selectionStatus: SelectionStatus.UNSELECTED,
-          calendar_id: calendarID,
-          availability: availability,
-          available_calendars: freeCalendars,
-          start_time: time,
-          end_time: endTime,
-          isBookable:
-            dateOffset >= 0 &&
-            dateOffset >= _this.min_book_ahead_days &&
-            dateOffset <= _this.max_book_ahead_days,
-        };
-      });
-  };
-
   let ticks: Date[] = [];
 
   const minimumTickHeight = 30; // minimum pixels between ticks, vertically
@@ -678,7 +456,10 @@
           numFreeCalendars === 0 ||
           fewerCalendarsThanRatio ||
           (_this.required_participants.length &&
-            !allRequiredParticipantsIncluded(epoch[0].available_calendars))
+            !arrayContainsArray(
+              requiredParticipants,
+              epoch[0].available_calendars,
+            ))
         ) {
           if (epoch[0].availability === AvailabilityStatus.CLOSED) {
             status = "closed";
@@ -767,7 +548,13 @@
   let days: Day[];
   $: days = dayRange.map((timestamp: Date) => {
     const slots = checkOverbooked(
-      generateDaySlots(timestamp, _this.start_hour, _this.end_hour),
+      generateDaySlots(
+        timestamp,
+        allCalendars,
+        calendarID,
+        requiredParticipants,
+        _this,
+      ),
     ); // TODO: include other potential post-all-slots-established checks, like overbooked, in a single secondary run here.
 
     return {
@@ -826,21 +613,6 @@
       });
       lastDispatchedSlots = sortedSlots;
     }
-  }
-
-  // https://derickbailey.com/2015/09/07/check-for-date-range-overlap-with-javascript-arrays-sorting-and-reducing/
-  function overlap(events: TimeSlot[], slot: TimeSlot) {
-    return events.reduce((result, current) => {
-      const overlap =
-        slot.start_time < current.end_time &&
-        current.start_time < slot.end_time;
-
-      if (overlap) {
-        // store the amount of overlap
-        result++;
-      }
-      return result;
-    }, 0);
   }
   // #endregion timeSlot selection
 
@@ -1616,7 +1388,7 @@
 
   //#region colours
   // Show partial availability as a gradient, rather than as categorically "partial"
-  $: partialScale = scaleLinear()
+  $: partialScale = scaleLinear<string>()
     .domain([0, allCalendars.length / 2, allCalendars.length])
     .range([_this.busy_color, _this.partial_color, _this.free_color]);
   //#endregion colours
