+++ package.json
@@ -51,6 +51,7 @@
     "@testing-library/cypress": "^8.0.2",
     "@testing-library/svelte": "^3.0.3",
     "@tsconfig/svelte": "^1.0.10",
+    "@types/d3-scale": "^4.0.2",
     "@types/d3-time": "^3.0.0",
     "@types/dompurify": "^2.3.1",
     "@types/fs-extra": "9.0.12",
